# Optimise `ColourInfo` conversion and add conversion helper method