--- a/osu.Framework/Graphics/Lines/Path_DrawNode.cs
+++ b/osu.Framework/Graphics/Lines/Path_DrawNode.cs
@@ -56,8 +56,8 @@ namespace osu.Framework.Graphics.Lines
 
             private Vector2 relativePosition(Vector2 localPos) => Vector2.Divide(localPos, drawSize);
 
-            private Color4 colourAt(Vector2 localPos) => DrawColourInfo.Colour.HasSingleColour
-                ? ((SRGBColour)DrawColourInfo.Colour).Linear
+            private Color4 colourAt(Vector2 localPos) => DrawColourInfo.Colour.TryExtractSingleColour(out SRGBColour colour)
+                ? colour.Linear
                 : DrawColourInfo.Colour.Interpolate(relativePosition(localPos)).Linear;
 
             private void addLineCap(Vector2 origin, float theta, float thetaDiff, RectangleF texRect)