The performance gains here are mostly theoretical and I can't benchmark them meaningfully in a real-world scenario. Consider this more of a refactoring that adds an optimal (and more user-friendly) method for checking whether a `ColourInfo` represents a single colour and then returning that colour.

This comes off the back of the osu!-side smoke implementation (where I initially noticed this), which potentially uses this method for tens of thousands of vertices (up to 30k is what I was able to achieve after [the recent changes](https://github.com/ppy/osu/pull/20827)). Similar is also done with slider paths (e.g. at some point https://osu.ppy.sh/beatmapsets/756794#osu/1605148 draws some 20k vertices). So _at most_ you'd see ~0.2ms frametime improvement.

Before
|              Method |     Mean |    Error |   StdDev | Allocated |
|-------------------- |---------:|---------:|---------:|----------:|
| ConvertToSRGBColour | 14.31 ns | 0.139 ns | 0.123 ns |         - |
|     ConvertToColor4 | 15.69 ns | 0.122 ns | 0.102 ns |         - |

After:
|                    Method |       Mean |     Error |    StdDev | Allocated |
|-------------------------- |-----------:|----------:|----------:|----------:|
|       ConvertToSRGBColour |  6.9628 ns | 0.0684 ns | 0.0534 ns |         - |
|           ConvertToColor4 | 11.1604 ns | 0.1658 ns | 0.1551 ns |         - |
| ExtractAndConvertToColor4 |  0.4039 ns | 0.0247 ns | 0.0231 ns |         - |

It uses concepts I've learnt over the years:
- Methods with `throw` are not inlined - extract these methods to promote inlining.
- Having strings in methods, even if the code is not hit, leads to a significant increase in ASM putting pressure on the JIT.

For comparisons sake, this is the best "real-world" example I could make, by looping `TestSceneSmoke` over 30 seconds:
| before | after |
| --- | --- |
| ![a](https://user-images.githubusercontent.com/1329837/197166090-243db5c7-d66b-442e-8e02-17fa333353e2.png) | ![b](https://user-images.githubusercontent.com/1329837/197166112-25df9044-cf26-4c6c-abda-d8283dd8352e.png) |

Again, I don't think these are indicative of anything so don't put much weight to them. There's too much noise.