@smoogipoo 2022-10-23
Fixed tests, the assertion was added post-fact.